fix: address code review feedback and improve test structure
- Replace json.Encoder with json.Marshal for idiomatic Go code
- Update GitHub Actions to latest versions (v4/v5/v6)
- Remove actions/cache step (integrated into setup-go v5)
- Add explicit build step and timeout to CI pipeline
- Add CI status badge to README
- Unify MockRepository across test files to improve maintainability

This addresses all Priority 1-3 review comments and implements
the recommended test mock consolidation for better code reuse.
